+++ src/agent.h
@@ -101,9 +101,9 @@ struct juice_agent {
 	ice_candidate_pair_t candidate_pairs[MAX_CANDIDATE_PAIRS_COUNT];
 	ice_candidate_pair_t *ordered_pairs[MAX_CANDIDATE_PAIRS_COUNT];
 	ice_candidate_pair_t *selected_pair;
-	size_t candidate_pairs_count;
+	int candidate_pairs_count;
 	agent_stun_entry_t entries[MAX_STUN_ENTRIES_COUNT];
-	size_t entries_count;
+	int entries_count;
 #ifdef NO_ATOMICS
 	agent_stun_entry_t *volatile selected_entry;
 #else
